Conflicts:
	model/model-impl/src/main/java/com/evolveum/midpoint/model/lens/AccountConstruction.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/lens/ChangeExecutor.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/lens/Clockwork.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/lens/LensContext.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/lens/LensDebugListener.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/lens/projector/Projector.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/lens/projector/UserPolicyProcessor.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/sync/CorrelationConfirmationEvaluator.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/sync/action/AddUserAction.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/sync/action/BaseAction.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/sync/action/DeleteUserAction.java
	model/model-impl/src/main/java/com/evolveum/midpoint/model/sync/action/SynchronizeAction.java
	model/model-impl/src/test/java/com/evolveum/midpoint/model/AbstractInternalModelIntegrationTest.java
	model/model-impl/src/test/java/com/evolveum/midpoint/model/lens/TestProjector.java
	model/model-intest/src/test/java/com/evolveum/midpoint/model/intest/sync/AbstractInboundSyncTest.java

+++ provisioning/provisioning-impl/src/main/java/com/evolveum/midpoint/provisioning/impl/ResourceObjectConverter.java
@@ -55,6 +55,7 @@ import com.evolveum.midpoint.prism.delta.ContainerDelta;
 import com.evolveum.midpoint.prism.delta.ItemDelta;
 import com.evolveum.midpoint.prism.delta.ObjectDelta;
 import com.evolveum.midpoint.prism.delta.PropertyDelta;
+import com.evolveum.midpoint.prism.match.MatchingRuleRegistry;
 import com.evolveum.midpoint.prism.path.ItemPath;
 import com.evolveum.midpoint.prism.query.AndFilter;
 import com.evolveum.midpoint.prism.query.EqualsFilter;
@@ -141,7 +142,10 @@ public class ResourceObjectConverter {
 	
 	@Autowired(required=true)
 	private EntitlementConverter entitlementConverter;
-	
+
+	@Autowired(required=true)
+	private MatchingRuleRegistry matchingRuleRegistry;
+
 	@Autowired(required=true)
 	private PrismContext prismContext;
 
@@ -154,10 +158,11 @@ public class ResourceObjectConverter {
 	
 	public PrismObject<ShadowType> getResourceObject(ConnectorInstance connector, ResourceType resource, 
 			Collection<? extends ResourceAttribute<?>> identifiers,
-			RefinedObjectClassDefinition objectClassDefinition, OperationResult parentResult) throws ObjectNotFoundException,
-			CommunicationException, SchemaException, ConfigurationException, SecurityViolationException, GenericConnectorException {
-
-		AttributesToReturn attributesToReturn = createAttributesToReturn(objectClassDefinition, identifiers, resource);
+			RefinedObjectClassDefinition objectClassDefinition, OperationResult parentResult)
+					throws ObjectNotFoundException, CommunicationException, SchemaException, ConfigurationException,
+					SecurityViolationException, GenericConnectorException {
+		
+		AttributesToReturn attributesToReturn = ProvisioningUtil.createAttributesToReturn(objectClassDefinition, resource);
 		
 		PrismObject<ShadowType> resourceShadow = fetchResourceObject(connector, resource, objectClassDefinition, identifiers, 
 				attributesToReturn, parentResult);
@@ -174,8 +179,10 @@ public class ResourceObjectConverter {
 			RefinedObjectClassDefinition objectClassDefinition, OperationResult parentResult) throws ObjectNotFoundException,
 			CommunicationException, SchemaException, ConfigurationException, SecurityViolationException, GenericConnectorException {
 
+		AttributesToReturn attributesToReturn = ProvisioningUtil.createAttributesToReturn(objectClassDefinition, resource);
+		
 		if (hasAllIdentifiers(identifiers, objectClassDefinition)) {
-			return getResourceObject(connector, resource, identifiers, objectClassDefinition, parentResult);
+			return fetchResourceObject(connector, resource, objectClassDefinition, identifiers, attributesToReturn, parentResult);
 		} else {
 			// Search
 			Collection<? extends RefinedAttributeDefinition> secondaryIdentifierDefs = objectClassDefinition.getSecondaryIdentifiers();
@@ -213,7 +220,7 @@ public class ResourceObjectConverter {
 				}
 			};
 			try {
-				connector.search(objectClassDefinition, query, handler, parentResult);
+				connector.search(objectClassDefinition, query, handler, attributesToReturn, parentResult);
 				if (shadowHolder.isEmpty()) {
 					throw new ObjectNotFoundException("No object found for secondary identifier "+secondaryIdentifier);
 				}
@@ -243,54 +250,7 @@ public class ResourceObjectConverter {
 		return true;
 	}
 
-	private AttributesToReturn createAttributesToReturn(RefinedObjectClassDefinition objectClassDefinition,
-			Collection<? extends ResourceAttribute<?>> identifiers, ResourceType resource) throws SchemaException {
-		boolean apply = false;
-		AttributesToReturn attributesToReturn = new AttributesToReturn();
-		attributesToReturn.setReturnDefaultAttributes(true);
-		
-		// Attributes
-		Collection<ResourceAttributeDefinition> explicit = new ArrayList<ResourceAttributeDefinition>();
-		for (RefinedAttributeDefinition attributeDefinition: objectClassDefinition.getAttributeDefinitions()) {
-			AttributeFetchStrategyType fetchStrategy = attributeDefinition.getFetchStrategy();
-			if (fetchStrategy != null && fetchStrategy == AttributeFetchStrategyType.EXPLICIT) {
-				explicit.add(attributeDefinition);
-			}
-		}
-		
-		if (!explicit.isEmpty()) {
-			attributesToReturn.setAttributesToReturn(explicit);
-			apply = true;
-		}
-		
-		// Password
-		CredentialsCapabilityType credentialsCapabilityType = ResourceTypeUtil.getEffectiveCapability(resource, CredentialsCapabilityType.class);
-		if (CapabilityUtil.isPasswordReturnedByDefault(credentialsCapabilityType)) {
-			// There resource is capable of returning password but it does not do it by default
-			AttributeFetchStrategyType passwordFetchStrategy = objectClassDefinition.getPasswordFetchStrategy();
-			if (passwordFetchStrategy == AttributeFetchStrategyType.EXPLICIT) {
-				attributesToReturn.setReturnPasswordExplicit(true);
-				apply = true;
-			}
-		}
-		
-		// Activation/administrativeStatus
-		ActivationCapabilityType activationCapabilityType = ResourceTypeUtil.getEffectiveCapability(resource, ActivationCapabilityType.class);
-		if (CapabilityUtil.isActivationStatusReturnedByDefault(activationCapabilityType)) {
-			// There resource is capable of returning enable flag but it does not do it by default
-			AttributeFetchStrategyType administrativeStatusFetchStrategy = objectClassDefinition.getActivationFetchStrategy(ActivationType.F_ADMINISTRATIVE_STATUS);
-			if (administrativeStatusFetchStrategy == AttributeFetchStrategyType.EXPLICIT) {
-				attributesToReturn.setReturnAdministrativeStatusExplicit(true);
-				apply = true;
-			}
-		}
-		
-		if (apply) {
-			return attributesToReturn;
-		} else {
-			return null;
-		}
-	}
+	
 
 	public PrismObject<ShadowType> addResourceObject(ConnectorInstance connector, ResourceType resource, 
 			PrismObject<ShadowType> shadow, RefinedObjectClassDefinition objectClassDefinition, OperationProvisioningScriptsType scripts, OperationResult parentResult)
@@ -322,6 +282,10 @@ public class ResourceObjectConverter {
 								SchemaDebugUtil.debugDump(additionalOperations,2) });
 			}
 			checkActivationAttribute(shadowType, resource, objectClassDefinition);
+			
+			if (!ResourceTypeUtil.hasCreateCapability(resource)){
+				throw new UnsupportedOperationException("Resource does not support 'create' operation");
+			}
 			
 			resourceAttributesAfterAdd = connector.addObject(shadowClone, additionalOperations, parentResult);
 
@@ -405,6 +369,10 @@ public class ResourceObjectConverter {
 								SchemaDebugUtil.debugDump(identifiers),
 								SchemaDebugUtil.debugDump(additionalOperations) });
 			}
+			
+			if (!ResourceTypeUtil.hasDeleteCapability(resource)){
+				throw new UnsupportedOperationException("Resource does not support 'delete' operation");
+			}
 
 			connector.deleteObject(objectClassDefinition, additionalOperations, identifiers, parentResult);
 
@@ -430,7 +398,7 @@ public class ResourceObjectConverter {
 	public Collection<PropertyModificationOperation> modifyResourceObject(
 			ConnectorInstance connector, ResourceType resource,
 			RefinedObjectClassDefinition objectClassDefinition, PrismObject<ShadowType> shadow, OperationProvisioningScriptsType scripts,
-			Collection<? extends ItemDelta> objectDeltas, OperationResult parentResult)
+			Collection<? extends ItemDelta> itemDeltas, OperationResult parentResult)
 			throws ObjectNotFoundException, SchemaException, CommunicationException, ConfigurationException,
 			SecurityViolationException, ObjectAlreadyExistsException {
 
@@ -440,13 +408,21 @@ public class ResourceObjectConverter {
 		Collection<? extends ResourceAttribute<?>> attributes = ShadowUtil.getAttributes(shadow);
 
 		if (isProtectedShadow(resource, objectClassDefinition, attributes)) {
-			LOGGER.error("Attempt to modify protected resource object " + objectClassDefinition + ": "
-					+ identifiers + "; ignoring the request");
-			throw new SecurityViolationException("Cannot modify protected resource object "
-					+ objectClassDefinition + ": " + identifiers);
+			if (hasChangesOnResource(itemDeltas)) {
+				LOGGER.error("Attempt to modify protected resource object " + objectClassDefinition + ": "
+						+ identifiers);
+				throw new SecurityViolationException("Cannot modify protected resource object "
+						+ objectClassDefinition + ": " + identifiers);
+			} else {
+				// Return immediately. This structure of the code makes sure that we do not execute any
+				// resource operation for protected account even if there is a bug in the code below.
+				LOGGER.trace("No resource modfications for protected resource object {}: {}; skipping",
+						objectClassDefinition, identifiers);
+				return null;
+			}
 		}
 		
-		collectAttributeAndEntitlementChanges(objectDeltas, operations, resource, shadow, objectClassDefinition);
+		collectAttributeAndEntitlementChanges(itemDeltas, operations, resource, shadow, objectClassDefinition);
 		
 		Collection<PropertyModificationOperation> sideEffectChanges = null;
 				
@@ -469,12 +445,12 @@ public class ResourceObjectConverter {
 		}
 		
 		// Execute entitlement modification on other objects (if needed)
-		executeEntitlementChangesModify(connector, resource, objectClassDefinition, shadow, scripts, objectDeltas, parentResult);
+		executeEntitlementChangesModify(connector, resource, objectClassDefinition, shadow, scripts, itemDeltas, parentResult);
 		
 		parentResult.recordSuccess();
 		return sideEffectChanges;
 	}
-	
+
 	private Collection<PropertyModificationOperation> executeModify(ConnectorInstance connector, ResourceType resource,
 			RefinedObjectClassDefinition objectClassDefinition, Collection<? extends ResourceAttribute<?>> identifiers, 
 					Collection<Operation> operations, OperationResult parentResult) throws ObjectNotFoundException, CommunicationException, SchemaException, SecurityViolationException, ConfigurationException, ObjectAlreadyExistsException {
@@ -521,6 +497,10 @@ public class ResourceObjectConverter {
 						"PROVISIONING MODIFY operation on {}\n MODIFY object, object class {}, identified by:\n{}\n changes:\n{}",
 						new Object[] { resource, PrettyPrinter.prettyPrint(objectClassDefinition.getTypeName()),
 								SchemaDebugUtil.debugDump(identifiers,1), SchemaDebugUtil.debugDump(operations,1) });
+			}
+			
+			if (!ResourceTypeUtil.hasUpdateCapability(resource)){
+				throw new UnsupportedOperationException("Resource does not support 'update' operation");
 			}
 			
 			sideEffectChanges = connector.modifyObject(objectClassDefinition, identifiers, operations,
@@ -642,6 +622,8 @@ public class ResourceObjectConverter {
 			final ResultHandler<ShadowType> resultHandler, ObjectQuery query, final OperationResult parentResult) throws SchemaException,
 			CommunicationException, ObjectNotFoundException, ConfigurationException {
 		
+		AttributesToReturn attributesToReturn = ProvisioningUtil.createAttributesToReturn(objectClassDef, resourceType);
+		
 		ResultHandler<ShadowType> innerResultHandler = new ResultHandler<ShadowType>() {
 			@Override
 			public boolean handle(PrismObject<ShadowType> shadow) {
@@ -659,7 +641,7 @@ public class ResourceObjectConverter {
 		};
 		
 		try {
-			connector.search(objectClassDef, query, innerResultHandler, parentResult);
+			connector.search(objectClassDef, query, innerResultHandler, attributesToReturn, parentResult);
 		} catch (GenericFrameworkException e) {
 			parentResult.recordFatalError("Generic error in the connector: " + e.getMessage(), e);
 			throw new SystemException("Generic error in the connector: " + e.getMessage(), e);
@@ -732,6 +714,11 @@ public class ResourceObjectConverter {
 			CommunicationException, SchemaException, SecurityViolationException, ConfigurationException {
 
 		try {
+			
+			if (!ResourceTypeUtil.hasReadCapability(resource)){
+				throw new UnsupportedOperationException("Resource does not support 'read' operation");
+			}
+			
 			PrismObject<ShadowType> resourceObject = connector.fetchObject(ShadowType.class, objectClassDefinition, identifiers,
 					attributesToReturn, parentResult);
 			return postProcessResourceObjectRead(connector, resource, resourceObject, objectClassDefinition, parentResult);
@@ -930,6 +917,21 @@ public class ResourceObjectConverter {
 			}
 		}		
 	}
+	
+	private boolean hasChangesOnResource(
+			Collection<? extends ItemDelta> itemDeltas) {
+		for (ItemDelta itemDelta : itemDeltas) {
+			if (new ItemPath(ShadowType.F_ATTRIBUTES).equals(itemDelta.getParentPath()) || SchemaConstants.PATH_PASSWORD.equals(itemDelta.getParentPath())) {
+				return true;
+			} else if (SchemaConstants.PATH_ACTIVATION.equals(itemDelta.getParentPath())){
+				return true;
+			} else if (new ItemPath(ShadowType.F_ASSOCIATION).equals(itemDelta.getPath())) { 
+				return true;				
+			}
+		}
+		return false;
+	}
+
 
 	private void collectAttributeAndEntitlementChanges(Collection<? extends ItemDelta> objectChange, 
 			Collection<Operation> operations, ResourceType resource, PrismObject<ShadowType> shadow, 
@@ -975,10 +977,12 @@ public class ResourceObjectConverter {
 		Validate.notNull(resource, "Resource must not be null.");
 		Validate.notNull(parentResult, "Operation result must not be null.");
 
-		LOGGER.trace("Shadow converter, START fetch changes");
+		LOGGER.trace("START fetch changes");
 
+		AttributesToReturn attrsToReturn = ProvisioningUtil.createAttributesToReturn(objectClass, resource);
+		
 		// get changes from the connector
-		List<Change<ShadowType>> changes = connector.fetchChanges(objectClass, lastToken, parentResult);
+		List<Change<ShadowType>> changes = connector.fetchChanges(objectClass, lastToken, attrsToReturn, parentResult);
 
 		Iterator<Change<ShadowType>> iterator = changes.iterator();
 		while (iterator.hasNext()) {
@@ -990,7 +994,7 @@ public class ResourceObjectConverter {
 					try {
 						
 						PrismObject<ShadowType> currentShadow = fetchResourceObject(connector, resource, objectClass, 
-								change.getIdentifiers(), null, parentResult);
+								change.getIdentifiers(), attrsToReturn, parentResult);
 						change.setCurrentShadow(currentShadow);
 						
 					} catch (ObjectNotFoundException ex) {
@@ -1011,7 +1015,7 @@ public class ResourceObjectConverter {
 		}
 
 		parentResult.recordSuccess();
-		LOGGER.trace("Shadow converter, END fetch changes");
+		LOGGER.trace("END fetch changes ({} changes)", changes == null ? "null" : changes.size());
 		return changes;
 	}
 	
@@ -1022,17 +1026,13 @@ public class ResourceObjectConverter {
 			PrismObject<ShadowType> resourceObject, RefinedObjectClassDefinition objectClassDefinition, OperationResult parentResult) throws SchemaException, CommunicationException, GenericFrameworkException {
 		
 		ShadowType resourceObjectType = resourceObject.asObjectable();
-		
-		// Protected object
-		if (isProtectedShadow(resourceType, resourceObject)) {
-			resourceObjectType.setProtectedObject(true);
-		}
+		setProtectedFlag(resourceType, resourceObject);
 		
 		// Simulated Activation
 		// FIXME??? when there are not native capabilities for activation, the
 		// resourceShadow.getActivation is null and the activation for the repo
 		// shadow are not completed..therefore there need to be one more check,
-		// we must chceck not only if the activation is null, but if it is, also
+		// we must check not only if the activation is null, but if it is, also
 		// if the shadow doesn't have defined simulated activation capability
 		if (resourceObjectType.getActivation() != null || ResourceTypeUtil.hasActivationCapability(resourceType)) {
 			ActivationType activationType = completeActivation(resourceObject, resourceType, parentResult);
@@ -1049,7 +1049,11 @@ public class ResourceObjectConverter {
 		return resourceObject;
 	}
 	
-	
+	public void setProtectedFlag(ResourceType resourceType, PrismObject<ShadowType> resourceObject) throws SchemaException {
+		if (isProtectedShadow(resourceType, resourceObject)) {
+			resourceObject.asObjectable().setProtectedObject(true);
+		}
+	}
 
 	/**
 	 * Completes activation state by determinig simulated activation if
@@ -1433,16 +1437,20 @@ public class ResourceObjectConverter {
 		// HACK FIXME
 		RefinedObjectClassDefinition refinedAccountDef = refinedSchema
 				.findRefinedDefinitionByObjectClassQName(ShadowKindType.ACCOUNT, objectClass);
-		LOGGER.trace("isProtectedShadow: {} -> {}, {}", new Object[] { objectClass, refinedAccountDef,
-				attributes });
+		boolean isProtected = false;
 		if (refinedAccountDef == null) {
-			return false;
+			isProtected = false;
+		} else {
+			Collection<ResourceObjectPattern> protectedAccountPatterns = refinedAccountDef.getProtectedObjectPatterns();
+			if (protectedAccountPatterns == null) {
+				isProtected = false;
+			} else {
+				isProtected = ResourceObjectPattern.matches(attributes, protectedAccountPatterns, matchingRuleRegistry);
+			}
 		}
-		Collection<ResourceObjectPattern> protectedAccountPatterns = refinedAccountDef.getProtectedObjectPatterns();
-		if (protectedAccountPatterns == null) {
-			return false;
-		}
-		return ResourceObjectPattern.matches(attributes, protectedAccountPatterns);
+		LOGGER.trace("isProtectedShadow: {} -> {}, {} = {}", new Object[] { objectClass, refinedAccountDef,
+				attributes, isProtected });
+		return isProtected;
 	}
 
 	private PrismObjectDefinition<ShadowType> getShadowTypeDef() {
